Move storage classes into a separate module, add documentation
