+++ tlm_adjoint/__init__.py
@@ -32,6 +32,7 @@ from .interface import *             # noqa: F401
 from .manager import *               # noqa: F401
 from .optimization import *          # noqa: F401
 from .overloaded_float import *      # noqa: F401
+from .storage import *               # noqa: F401
 from .tangent_linear import *        # noqa: F401
 from .tlm_adjoint import *           # noqa: F401
 from .verification import *          # noqa: F401
